+++ csatar-plugins/csatar/classes/CsatarPivot.php
@@ -22,4 +22,11 @@ class CsatarPivot extends Pivot
         return $this->otherKey;
     }
 
+    public function getParent($id = null) {
+        return $this->parent->id ? $this->parent : $this->getParentById($id);
+    }
+
+    public function getParentById($id) {
+        return $this->getParentClass()::find($id);
+    }
 }