+++ csatar-plugins/csatar/models/Association.php
@@ -177,7 +177,7 @@ class Association extends OrganizationBase
     public function updateCache(): void
     {
         if ($this->wasRecentlyCreated) {
-            StructureTree::updateAssociationTree($this->association_id);
+            StructureTree::updateAssociationTree($this->id);
         }
 
         if (empty($this->original)) {
@@ -261,4 +261,11 @@ class Association extends OrganizationBase
         ];
     }
 
+    public function getSpecialWorkplanAgeGroupIdOptions() {
+        if (empty($this->id)) {
+            return [];
+        }
+
+        return AgeGroup::where('association_id', $this->id)->lists('name', 'id');
+    }
 }