+++ csatar-plugins/knowledgerepository/models/Game.php
@@ -132,6 +132,13 @@ class Game extends PermissionBasedAccess
         ],
     ];
 
+    public $morphOne = [
+        'photo' => [
+            \Csatar\KnowledgeRepository\Models\ActivityType::class,
+            'name' => 'programmable',
+        ],
+    ];
+
     public $attachMany = [
         'attachements' => ['System\Models\File'],
     ];